Make using DiskCache optional.

+++ CADETProcess/optimization/progress.py
@@ -91,6 +91,7 @@ class OptimizationProgress():
             working_directory,
             results_directory,
             save_results=False,
+            use_diskcache=True,
             cache_directory=None,
             keep_cache=True,
             overwrite=True):
@@ -118,11 +119,11 @@ class OptimizationProgress():
             self.progress_directory = None
 
         self.keep_cache = keep_cache
-        if self.save_results and cache_directory is None:
-            cache_directory = self.working_directory / 'cache'
-            cache_directory.mkdir(exist_ok=overwrite)
-        self.cache_directory = cache_directory
-        self.cache = ResultsCache(self.cache_directory)
+        if use_diskcache:
+            if self.save_results and cache_directory is None:
+                cache_directory = self.working_directory / 'cache'
+                cache_directory.mkdir(exist_ok=overwrite)
+        self.cache = ResultsCache(use_diskcache, cache_directory)
 
         if self.save_results:
             self.setup_figures()
@@ -626,26 +627,39 @@ class ResultsCache():
 
     """
 
-    def __init__(self, directory=None):
-        if directory is None:
-            directory = tempfile.mkdtemp(prefix='diskcache-')
-        self.directory = directory
-
-        self.cache = Cache(
-           directory,
-           disk=DillDisk,
-           disk_min_file_size=2**18,
-           size_limit=2**36,
-        )
+    def __init__(self, use_diskcache=True, directory=None):
+        self.init_cache(use_diskcache, directory)
 
         self.tags = defaultdict(list)
+
+    def init_cache(self, use_diskcache, directory):
+        if use_diskcache:
+            if directory is None:
+                directory = tempfile.mkdtemp(prefix='diskcache-')
+            self.directory = directory
+
+            self.cache = Cache(
+               directory,
+               disk=DillDisk,
+               disk_min_file_size=2**18,
+               size_limit=2**36,
+            )
+            self.directory = self.cache.directory
+        else:
+            self.cache = {}
+            self.directory = None
+
+        self.use_diskcache = use_diskcache
 
     def set(self, eval_obj, step, x, result, tag=None):
         key = f'{eval_obj}.{step}.{x}'
         if tag is not None:
             self.tags[tag].append(key)
 
-        self.cache.set(key, result, expire=None)
+        if self.use_diskcache:
+            self.cache.set(key, result, expire=None)
+        else:
+            self.cache[key] = result
 
     def get(self, eval_obj, step, x):
         key = f'{eval_obj}.{step}.{x}'
@@ -657,23 +671,33 @@ class ResultsCache():
     def delete(self, eval_obj, step, x):
         key = f'{eval_obj}.{step}.{x}'
 
-        self.cache.delete(key)
+        if self.use_diskcache:
+            self.cache.delete(key)
+        else:
+            self.cache.pop(key)
 
     def prune(self, tag='temp'):
         try:
             keys = self.tags.pop(tag)
             for key in keys:
-                self.cache.delete(key)
+                eval_obj, step, x = key.split('.')
+                self.delete(eval_obj, step, x)
         except KeyError:
             pass
 
     def close(self):
-        self.cache.close()
+        if self.use_diskcache:
+            self.cache.close()
 
-    def delete_database(self):
-        self.close()
-        try:
-            shutil.rmtree(self.directory, ignore_errors=True)
-        except FileNotFoundError:
-            pass
+    def delete_database(self, reinit=False):
+        if self.use_diskcache:
+            self.close()
+            try:
+                shutil.rmtree(self.directory, ignore_errors=True)
+            except FileNotFoundError:
+                pass
+
         self.cache = None
+
+        if reinit:
+            self.init_cache(self.use_diskcache, self.directory)
